Moved components to common folder
