+++ frontend/src/components/common/ConfirmDeleteDialog.tsx
