+++ frontend/src/components/pages/program-details/ProgramDetails.tsx
@@ -6,9 +6,7 @@ import { Box, TextField, Button, CircularProgress, styled } from '@mui/material'
 import type { Program } from 'types/Program'
 import { useParams, useNavigate } from 'react-router-dom'
 import { useAuth0 } from '@auth0/auth0-react'
-import { Footer } from '../../common'
-import { ConfirmDeleteDialog } from './components/ConfirmDeleteDialog'
-import { ExerciseFields } from './components/ExerciseFields'
+import { Footer, ExerciseFields, ConfirmDeleteDialog } from '../../common'
 import { useSnackbar } from 'context/SnackbarContext'
 import BACKEND_URL from '../../../constants'
