--- frontend/src/components/pages/program-details/components/index.ts
@@ -1,4 +1,0 @@
-import { ConfirmDeleteDialog } from './ConfirmDeleteDialog'
-import { ExerciseFields } from './ExerciseFields'
-
-export { ConfirmDeleteDialog, ExerciseFields }
