+++ frontend/src/demo/pages/demo-planner/components/index.ts
@@ -1,0 +1,3 @@
+import { DemoProgramList } from './DemoProgramList'
+
+export { DemoProgramList }
